mollex: change filtering order
Primarily done since the copy constructor of cv::Mat does not actually
copy, breaking the constant references. We don't want to modify
the image that was passed in. The easiest way is to change the filtering
order (which apparently doesn't really make a difference) because
the bilateral filter can't work in-place.