Add correct per-mollusc alpha export
Previously, we would use the same alpha mask for every mollusc,
meaning that if the bounding boxes of two molluscs intersected,
a tiny slice of another mollusc would be shown within the exported
image.
Furthermore, holes that the heuristic determined are not molluscs
were erroneously included in the exported alpha mask.

Both issues are fixed by exporting a per-contour alpha mask.